+++ grafana/grafana_tool.py
@@ -33,7 +33,7 @@ import urllib.parse
 import yaml
 
 import pydantic_utils
-import secret_utils
+import db_utils
 
 SQLALCHEMY_DB_SCHEME = "postgresql"
 DEFAULT_URL = "http://localhost:3000/grafana"
@@ -102,7 +102,7 @@ def do_create_db(args: Any) -> None:
 
     try:
         grafana_dsn = \
-            secret_utils.substitute_password(
+            db_utils.substitute_password(
                 dsc="Grafana database", dsn=settings.grafana_db_dsn,
                 password_file=settings.grafana_db_password_file)
     except RuntimeError as ex:
