+++ src/afc-packages/pkgs.als_siphon
@@ -1,2 +1,2 @@
-/wd/afc-packages/secret_utils
+/wd/afc-packages/db_utils
 /wd/afc-packages/defs
