+++ src/afc-packages/pkgs.cert_db
@@ -8,5 +8,5 @@
 /wd/afc-packages/als
 /wd/afc-packages/rcache
 /wd/afc-packages/healthchecks
-/wd/afc-packages/secret_utils
+/wd/afc-packages/db_utils
 /wd/afc-packages/prometheus_utils
