+++ src/afc-packages/pkgs.grafana
@@ -4,4 +4,4 @@
 # This work is licensed under the OpenAFC Project License, a copy of which is
 # included with this software program.
 /wd/afc-packages/pydantic_utils
-/wd/afc-packages/secret_utils
+/wd/afc-packages/db_utils
