+++ src/afc-packages/pkgs.worker
@@ -6,4 +6,4 @@
 /wd/afc-packages/netcli
 /wd/afc-packages/rcache
 /wd/afc-packages/als
-/wd/afc-packages/secret_utils
+/wd/afc-packages/db_utils
