+++ src/afc-packages/rcache/rcache_rmq.py
@@ -18,7 +18,7 @@ from typing import cast, List, Optional, Set
 
 from log_utils import error, get_module_logger
 from rcache_models import RCACHE_RMQ_EXCHANGE_NAME, RmqReqRespKey
-import secret_utils
+import db_utils
 
 __all__ = ["RcacheRmq", "RcacheRmqConnection"]
 
@@ -187,7 +187,7 @@ class RcacheRmq:
         try:
             self._url_params = pika.URLParameters(self.rmq_dsn)
         except pika.exceptions.AMQPError as ex:
-            error(f"RabbitMQ URL '{secret_utils.safe_dsn(self.rmq_dsn)}' has "
+            error(f"RabbitMQ URL '{db_utils.safe_dsn(self.rmq_dsn)}' has "
                   f"invalid syntax: {ex}")
 
     def create_connection(self, tx_queue_name: Optional[str] = None) \
