+++ uls/uls_service_healthcheck.py
@@ -24,7 +24,7 @@ import sys
 from typing import Any, cast, Dict, List, NamedTuple, Optional, Set, Union
 
 from pydantic_utils import env_help, merge_args
-from secret_utils import safe_dsn
+from db_utils import safe_dsn
 from uls_service_common import *
 from uls_service_state_db import AlarmType, CheckType, DownloaderMilestone, \
     LogType, StateDb
